Create README.md
If you are new to *github* and was sent this webpage for any help regarding installations and commands, just open the first link and it'll open a home page. The home page displays important links regarding all installations and commands. GitHub can be a bit overwhelming at first, so refer to this mini-site until you are aquainted with the app.